Harden exec-scoped proxy attribution
